moved main to Command class, dropped generated setProperty, generated code formatted
